hw06/pyproject.toml
[tool.poetry]
name = "syn-tg-bot-hw06"
version = "0.1.0"
description = ""
authors = ["Michael Savin <jtprogru@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.8"
gunicorn = "^21.2.0"
uvicorn = "^0.27.0"
prometheus-client = "^0.19.0"
aiohttp = "^3.9.3"

[tool.isort]
line_length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"